New command registration idea (CommandRegistrator)
